Initial commit: LVGL on STM32H743ZI2
I managed to properly fix the rotations and orientations, contrary to the repos
I forked this from. We miss the touch support but we might need an RTOS for
that.

Also we can add some optimisations like better buffering and DMA transmissions
but these might not be worth it considering this display is low quality enough.

Signed-off-by: Daniel Vásquez <danielgusvt@gmail.com>